+++ pai/api/workspace.py
@@ -3,7 +3,17 @@ from __future__ import absolute_import
 import six
 
 from pai.api.base import paginate_service_call, BaseClient, BaseTeaClient
-from pai.libs.alibabacloud_aiworkspace20210204.models import ListWorkspacesRequest
+from pai.libs.alibabacloud_aiworkspace20210204.models import (
+    ListWorkspacesRequest,
+    ListResourcesRequest,
+    ListMembersRequest,
+    ListProductsRequest,
+    ListOperationLogsRequest,
+    CreateWorkspaceRequest,
+    CreateWorkspaceResourceRequest,
+    CreateMemberRequest,
+    DeleteMembersRequest,
+)
 
 
 class WorkspaceClient(BaseTeaClient):
@@ -13,7 +23,7 @@ class WorkspaceClient(BaseTeaClient):
     def __init__(self, base_client):
         super(WorkspaceClient, self).__init__(base_client=base_client)
 
-    def list_workspaces(
+    def list_workspace(
         self,
         page_number=None,
         page_size=None,
@@ -27,7 +37,7 @@ class WorkspaceClient(BaseTeaClient):
     ):
         request = ListWorkspacesRequest(
             page_number=page_number,
-            page_size=None,
+            page_size=page_size,
             sort_by=sort_by,
             order=order,
             workspace_name=workspace_name,
@@ -36,7 +46,37 @@ class WorkspaceClient(BaseTeaClient):
             option=option,
             verbose=verbose,
         )
-        self.base_client.list_workspaces(request)
+
+        resp = self._call_service_with_exception(
+            self.base_client.list_workspaces, request=request
+        ).to_map()
+
+        workspaces, total_count = resp["Workspaces"], resp["TotalCount"]
+        return workspaces, total_count
+
+    def list_workspace_generator(
+        self,
+        page_number=None,
+        page_size=None,
+        sort_by=None,
+        order=None,
+        workspace_name=None,
+        module_list=None,
+        status=None,
+        option=None,
+        verbose=None,
+    ):
+        return type(self).to_generator(self.list_workspace)(
+            page_number=page_number,
+            page_size=page_size,
+            sort_by=sort_by,
+            order=order,
+            workspace_name=workspace_name,
+            module_list=module_list,
+            status=status,
+            option=option,
+            verbose=verbose,
+        )
 
     def _get_endpoint(self):
         if self._endpoint:
@@ -51,108 +91,58 @@ class WorkspaceClient(BaseTeaClient):
     def _get_product(self):
         return "AIWorkSpace"
 
-    def create(self, name, alias=None, description=None):
-        request = self._construct_request(CreateWorkspaceRequest.CreateWorkspaceRequest)
-        request.set_WorkspaceName(name)
-        if alias:
-            request.set_WorkspaceAlias(alias)
-        if description:
-            request.set_Description(description)
-        return self._call_service_with_exception(request)
+    def create(self, name, display_name=None, description=None, env_types=None):
+        request = CreateWorkspaceRequest(
+            workspace_name=name,
+            description=description,
+            display_name=display_name,
+            env_types=None,
+        )
+
+        resp = self._call_service_with_exception(
+            self.base_client.create_workspace, request=request
+        )
+        return resp.to_map()
 
     def get_workspace(self, workspace_id):
         resp = self.base_client.get_workspace(workspace_id)
         return resp.body.to_map()
 
     def update(self, workspace_id, name):
-        request = self._construct_request(UpdateWorkspaceRequest.UpdateWorkspaceRequest)
-        request.set_WorkspaceId(workspace_id)
-        request.set_WorkspaceName(name)
-        return self._call_service_with_exception(request)
+        raise NotImplementedError
 
     def delete_member(self, workspace_id, delete_member_ids):
-        if not delete_member_ids:
-            raise ValueError("Please given non-empty delete_member_ids")
-
-        request = self._construct_request(UpdateWorkspaceRequest.UpdateWorkspaceRequest)
-        if not isinstance(delete_member_ids, (list, tuple)):
-            delete_member_ids = [delete_member_ids]
-
-        request.set_WorkspaceId(workspace_id)
-        request.add_body_params("DeletedMemberIds", delete_member_ids)
-        return self._call_service_with_exception(request)
+        raise NotImplementedError
 
     def add_member(self, workspace_id, member):
-        request = self._construct_request(CreateMemberRequest.CreateMemberRequest)
-        if not isinstance(member, (list, tuple)):
-            member = [member]
+        raise NotImplementedError
 
-        request.set_WorkspaceId(workspace_id)
-        request.add_body_params("Members", member)
-
-        return self._call_service_with_exception(request)
-
-    @paginate_service_call
     def list_member(self, workspace_id, name=None, role=None):
-        request = self._construct_request(ListMembersRequest.ListMembersRequest)
-        request.set_WorkspaceId(workspace_id)
-        if name is not None:
-            request.set_UserName(name)
-
-        if role is not None and role:
-            if isinstance(role, (list, tuple)):
-                role = ",".join(role)
-            elif not isinstance(role, six.string_types):
-                raise ValueError("Required comma split roles list")
-            request.set_Roles(role)
-        return request
+        raise NotImplementedError
 
     def list_sub_users(self, exclude_workspace_id=None):
-        request = self._construct_request(ListSubUsersRequest.ListSubUsersRequest)
-        if exclude_workspace_id:
-            request.set_ExcludeWorkspaceId(exclude_workspace_id)
-        return self._call_service_with_exception(request)
+        raise NotImplementedError
 
     def list_permissions(self, workspace_id):
-        request = self._construct_request(ListPermissionsRequest.ListPermissionsRequest)
-        request.set_WorkspaceId(workspace_id)
-        return self._call_service_with_exception(request)
+        raise NotImplementedError
 
     def get_permission(self, workspace_id, permission_code):
-        request = self._construct_request(GetPermissionRequest.GetPermissionRequest)
-        request.set_WorkspaceId(workspace_id)
-        request.set_PermissionCode(permission_code)
-        return self._call_service_with_exception(request)
+        raise NotImplementedError
 
     def create_tenant(self):
-        request = self._construct_request(CreateTenantRequest.CreateTenantRequest)
-        return self._call_service_with_exception(request)
+        raise NotImplementedError
 
     def get_tenant(self):
-        request = self._construct_request(GetTenantRequest.GetTenantRequest)
-        return self._call_service_with_exception(request)
+        raise NotImplementedError
 
     def add_compute_engine(self, workspace_id):
-        pass
+        raise NotImplementedError
 
-    @paginate_service_call
     def list_compute_engines(self, name, workspace_id, engine_type="PAI", **kwargs):
-        request = self._construct_request(ListResourcesRequest.ListResourcesRequest)
-        if name:
-            request.set_ResourceName(name)
-        if workspace_id:
-            request.set_WorkspaceId(str(workspace_id))
-        if engine_type:
-            request.set_ProductType(engine_type)
-        return request
+        raise NotImplementedError
 
     def list_resource_groups(self, tenant_id):
-        request = self._construct_request(
-            ListResourceGroupsRequest.ListResourceGroupsRequest
-        )
-        request.set_TenantId(tenant_id)
-        return self._call_service_with_exception(request)
+        raise NotImplementedError
 
     def list_commodities(self):
-        request = self._construct_request(ListCommoditiesRequest.ListCommoditiesRequest)
-        return self._call_service_with_exception(request)
+        raise NotImplementedError